Add prelim support to personacrypt for PEFS+SYNCTHING
This will allow setting up a PEFS encrypted home-directory along with
a SyncThing target of the encrypted contents. When exporting and importing
the user setting, the SyncThing Client ID's and folder ID's will be
preserved, alowing the linking of a users $HOME directory between more
than one system.

Changelog: YES